Initial version of ci tracing (#41)
Add a first prototype of tracing to the velo-action. There is a bunch
of stuff left to implement in follow up PR's such as how to deal with
the case where this action runs multiple times in one workflow and
how to deal with more complicated workflow graphs. Currently it
only support a two workflow layout where CI -> Deploy.

This is basically how it works in the example project:
https://github.com/kolonialno/example-deploy-project/commit/eaee3741bf615315ac2527f04c119f8c1b6c5a96