Fix release notes and default to sha version
The release notes currently show the wrong sha occasionally, the reason
for this behaviour is defaulting to the HEAD of main on the remote as
explained here:
https://github.community/t/github-sha-not-the-same-as-the-triggering-commit/18286/3

This commit fixes that behaviour. In addition we want to default to
short git_sha's as the version by default. Since this change ties in
nicely with that I decided to do it in one go.